requirements.txt
paho-mqtt
pymongo
pynput